Falta Cosillas

+++ Coordinate-Times/src/main/java/controlador/perfilC.java
@@ -17,6 +17,7 @@ public class perfilC{
 
         private Integer usuario;
         private Base u = new Base();
+        private UsuarioBean ub = new UsuarioBean();
 /**
  * Método constructor de la clase perfilC
  * 
@@ -31,29 +32,16 @@ public class perfilC{
          * @return null si el registro es correcto y un mensaje de error en otro caso
          */
 	public String eliminaPerfil(){
-               Usuario condicion = UsuarioBean.getUsuario1();
-                if(UsuarioBean.getUsuario1() == null){
-                    FacesContext.getCurrentInstance().addMessage(null
-                                                         , new FacesMessage(FacesMessage.SEVERITY_ERROR, "Inicia sesion para poder acceder a esta funcion", ""));
-
-                }else{
-                usuario = UsuarioBean.getUsuario1().getTipo();		
-                if(usuario != 0){
-			FacesContext.getCurrentInstance().addMessage(null
-                                                         , new FacesMessage(FacesMessage.SEVERITY_ERROR, "No tienes los permisos necesarios", ""));
-		}else{
-                 if(condicion.getCorreo()==(BuscarPorPerfil.getUs().get(0)).getCorreo()){
+          if(ub.esAdministrador()){
                  FacesContext.getCurrentInstance().addMessage(null,
-                    new FacesMessage(FacesMessage.SEVERITY_ERROR, "No puedes eliminarte a ti mismo", ""));        
+                    new FacesMessage(FacesMessage.SEVERITY_ERROR, "No puedes eliminarte de esta forma Administrador, tendras que eliminarte directamente de la base", ""));        
                 }else{
 			FacesContext.getCurrentInstance().addMessage(null,
                     new FacesMessage(FacesMessage.SEVERITY_INFO, "Se ha eliminado correctamenten el perfil", ""));
                         u.eliminarU(BuscarPorPerfil.getUs().get(0));
                         usuario = null;
 		}
-                }
-                }
+                
         return null;
 	}
-
 }
